WI 705990 Rename packages for ECSP release
