--- src/main/java/com/harman/ignite/sql/IgniteSqlDaoApplication.java
@@ -1,59 +1,0 @@
-/*
- *
- *
- *   ******************************************************************************
- *
- *    Copyright (c) 2023-24 Harman International
- *
- *
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *
- *    you may not use this file except in compliance with the License.
- *
- *    You may obtain a copy of the License at
- *
- *
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *
- *    Unless required by applicable law or agreed to in writing, software
- *
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *
- *    See the License for the specific language governing permissions and
- *
- *    limitations under the License.
- *
- *
- *
- *    SPDX-License-Identifier: Apache-2.0
- *
- *    *******************************************************************************
- *
- *
- */
-
-package com.harman.ignite.sql;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.stereotype.Component;
-
-/**
- * Main component to load all the Spring components into the context of the implementing service. <br>
- * <br>
- * It allows scanning all the classes inside org.eclipse.ecsp packages
- * including the dependent libraries like utils and allowing loading of the
- * beans from the dependencies.
- *
- */
-@ComponentScan(basePackages = {"org.eclipse.ecsp"})
-@Component
-@Configuration
-public class IgniteSqlDaoApplication {
-
-}
