+++ src/main/java/org/eclipse/ecsp/sql/exception/SqlDaoException.java
@@ -1,0 +1,56 @@
+/*
+ *
+ *
+ *   ******************************************************************************
+ *
+ *    Copyright (c) 2023-24 Harman International
+ *
+ *
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *
+ *    you may not use this file except in compliance with the License.
+ *
+ *    You may obtain a copy of the License at
+ *
+ *
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ *    See the License for the specific language governing permissions and
+ *
+ *    limitations under the License.
+ *
+ *
+ *
+ *    SPDX-License-Identifier: Apache-2.0
+ *
+ *    *******************************************************************************
+ *
+ *
+ */
+
+package org.eclipse.ecsp.sql.exception;
+
+/**
+ * Exception class for SQL DAO.
+ */
+public class SqlDaoException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+
+    public SqlDaoException(String message) {
+        super(message);
+    }
+
+    public SqlDaoException(String message, Throwable throwable) {
+        super(message, throwable);
+    }
+}
